requirements.txt
openai>=1.55.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
Pillow>=10.0.0
chromadb>=0.4.0
pypdf>=4.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
